refactor: add styles for the card
