+++ src/components/Card/Card.props.ts
@@ -3,4 +3,13 @@ interface CardProps {
   description: string;
   onClick: () => void;
 }
+interface Pokemon {
+  name: string;
+  url: string;
+  height: number;
+  weight: number;
+  base_experience: number;
+}
 export default CardProps;
+
+export type { Pokemon };
