+++ src/components/DetailedCard/DetailedCard.props.ts
@@ -1,8 +1,8 @@
-import { Pokemon } from '../../pages/HomePage/HomePage.props';
-
+import { Pokemon } from '../Card/Card.props';
 interface DetailedCardProps {
   selectedItem: Pokemon;
   loading: boolean;
   onClose: () => void;
 }
+
 export default DetailedCardProps;
