+++ src/index.css
@@ -2,19 +2,18 @@
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
   font-weight: 400;
-display: flex;
-justify-content: center;
-width: 100%;
+  display: flex;
+  justify-content: center;
+  width: 100%;
   color-scheme: light dark;
   color: rgba(255, 255, 255, 0.87);
   background-color: #7ba5f3;
-
   font-synthesis: none;
   text-rendering: optimizeLegibility;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
-
+#root{ width: 100%;}
 a {
   font-weight: 500;
   color: #646cff;
@@ -27,6 +26,7 @@ a:hover {
 body {
   margin: 0;
   display: flex;
+  width: 100%;
 }
 
 h1 {
