+++ src/pages/404.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const Page404 = () => {
   return (
     <div>
