+++ src/pages/HomePage/HomePage.module.css
@@ -1,22 +1,34 @@
 .containerHomePage {
     margin: 0 auto;
     width: 100%;
+    max-width: 1200px;
     height: 100vh;
     text-align: center;
   }
   
-  .leftSection {
-    width: 45%;
-    padding: 20px;
-  }
-  
-  .rightSection {
-    width: 45%;
-    padding: 20px;
-    background-color: #f5f5f5;
-  }
   .resultsWrapper {
     display: flex;
     justify-content: space-between;
     margin-top: 20px;
+    width: 100%;
+    min-height: 500px;
+    gap: 20px;
+  }
+  .leftSection {
+    width: 100%;
+    max-width: 60%;
+    background-color: #5b60f5;
+    border-radius: 12px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+    margin-bottom: 15px;
+    padding: 20px;
+  }
+  
+  .rightSection {
+    width: 25%;
+    background-color: #5b60f5;
+    border-radius: 12px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+    margin-bottom: 15px;
+    padding: 20px;
   }
